readings/requirements.txt
ads1015==0.0.5
astral==2.1
boto3==1.4.4
botocore==1.5.95
cbor2==4.1.2
cffi==1.14.0
click==6.6
docutils==0.16
enviroplus==0.0.3
font-roboto==0.0.1
fonts==0.0.3
greengrasssdk==1.5.0
i2cdevice==0.0.6
jmespath==0.9.5
ltr559==0.1.0
pimoroni-bme280==0.0.2
pms5003==0.0.5
pycparser==2.20
pyserial==3.4
python-dateutil==2.8.1
python-lambda==11.7.1
pytz==2020.1
PyYAML==5.1
RPi.GPIO==0.7.0
s3transfer==0.1.13
six==1.14.0
smbus==1.1.post2
smbus2==0.3.0
sounddevice==0.3.15
spidev==3.4
ST7735==0.0.4
